增加之前的界面

+++ app/src/main/java/com/loan/stl/utils/device/DeviceUtil.java
@@ -41,7 +41,7 @@ public class DeviceUtil {
      *
      * @return
      */
-    public static int dp2px(Context context, float dpValue) {
+    public static int dp2px(Context context, int dpValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
         return (int) (dpValue * scale + 0.5f);
     }
